pyproject.toml
[tool.poetry]
name = "iotd"
version = "0.1.0"
description = ""
authors = ["ellisajung <ellisajung12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
pillow = "^9.5.0"
django-nextjs = "^2.2.2"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
coverage = "^7.2.6"
pytest-cov = "^4.1.0"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
